Remove storage facade from ccompat
