+++ app/src/main/java/io/apicurio/registry/ccompat/rest/v7/impl/ConfigResourceImpl.java
@@ -31,8 +31,8 @@ import io.apicurio.registry.rules.compatibility.CompatibilityLevel;
 import io.apicurio.registry.storage.RuleNotFoundException;
 import io.apicurio.registry.storage.dto.RuleConfigurationDto;
 import io.apicurio.registry.types.RuleType;
-
 import jakarta.interceptor.Interceptors;
+
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
@@ -79,8 +79,7 @@ public class ConfigResourceImpl extends AbstractResource implements ConfigResour
     @Override
     @Authorized(style = AuthorizedStyle.None, level = AuthorizedLevel.Admin)
     public CompatibilityLevelParamDto getGlobalCompatibilityLevel() {
-        return getCompatibilityLevel(() ->
-                facade.getGlobalRule(RuleType.COMPATIBILITY).getConfiguration());
+        return getCompatibilityLevel(() -> storage.getGlobalRule(RuleType.COMPATIBILITY).getConfiguration());
     }
 
     @Override
@@ -88,8 +87,14 @@ public class ConfigResourceImpl extends AbstractResource implements ConfigResour
     @Authorized(style = AuthorizedStyle.None, level = AuthorizedLevel.Admin)
     public CompatibilityLevelDto updateGlobalCompatibilityLevel(CompatibilityLevelDto request) {
         updateCompatibilityLevel(request.getCompatibility(),
-                dto -> facade.createOrUpdateGlobalRule(RuleType.COMPATIBILITY, dto),
-                () -> facade.deleteGlobalRule(RuleType.COMPATIBILITY));
+                dto -> {
+                    if (!doesGlobalRuleExist(RuleType.COMPATIBILITY)) {
+                        storage.createGlobalRule(RuleType.COMPATIBILITY, dto);
+                    } else {
+                        storage.updateGlobalRule(RuleType.COMPATIBILITY, dto);
+                    }
+                },
+                () -> storage.deleteGlobalRule(RuleType.COMPATIBILITY));
         return request;
     }
 
@@ -98,16 +103,27 @@ public class ConfigResourceImpl extends AbstractResource implements ConfigResour
     @Authorized(style = AuthorizedStyle.ArtifactOnly, level = AuthorizedLevel.Write)
     public CompatibilityLevelDto updateSubjectCompatibilityLevel(String subject, CompatibilityLevelDto request, String groupId) {
         updateCompatibilityLevel(request.getCompatibility(),
-                dto -> facade.createOrUpdateArtifactRule(subject, RuleType.COMPATIBILITY, dto, groupId),
-                () -> facade.deleteArtifactRule(subject, RuleType.COMPATIBILITY, groupId));
+                dto -> {
+                    if (!doesArtifactRuleExist(subject, RuleType.COMPATIBILITY, groupId)) {
+                        storage.createArtifactRule(groupId, subject, RuleType.COMPATIBILITY, dto);
+                    } else {
+                        storage.updateArtifactRule(groupId, subject, RuleType.COMPATIBILITY, dto);
+                    }
+                },
+                () -> {
+                    try {
+                        storage.deleteArtifactRule(groupId, subject, RuleType.COMPATIBILITY);
+                    } catch (RuleNotFoundException e) {
+                        //Ignore, fail only when the artifact is not found
+                    }
+                });
         return request;
     }
 
     @Override
     @Authorized(style = AuthorizedStyle.ArtifactOnly, level = AuthorizedLevel.Read)
     public CompatibilityLevelParamDto getSubjectCompatibilityLevel(String subject, String groupId) {
-        return getCompatibilityLevel(() ->
-                facade.getArtifactRule(subject, RuleType.COMPATIBILITY, groupId).getConfiguration());
+        return getCompatibilityLevel(() -> storage.getArtifactRule(groupId, subject, RuleType.COMPATIBILITY).getConfiguration());
     }
 
     @Override
@@ -115,9 +131,9 @@ public class ConfigResourceImpl extends AbstractResource implements ConfigResour
     @Authorized(style = AuthorizedStyle.ArtifactOnly, level = AuthorizedLevel.Write)
     public CompatibilityLevelParamDto deleteSubjectCompatibility(String subject, String groupId) {
         final CompatibilityLevelParamDto compatibilityLevel = getCompatibilityLevel(() ->
-                facade.getArtifactRule(subject, RuleType.COMPATIBILITY, groupId).getConfiguration());
+                storage.getArtifactRule(groupId, subject, RuleType.COMPATIBILITY).getConfiguration());
         if (!CompatibilityLevel.NONE.name().equals(compatibilityLevel.getCompatibilityLevel())) {
-            facade.deleteArtifactRule(subject, RuleType.COMPATIBILITY, groupId);
+            storage.deleteArtifactRule(groupId, subject, RuleType.COMPATIBILITY);
         }
         return compatibilityLevel;
     }
