+++ app/src/main/java/io/apicurio/registry/ccompat/rest/v7/impl/SchemasResourceImpl.java
@@ -23,13 +23,24 @@ import io.apicurio.registry.auth.AuthorizedStyle;
 import io.apicurio.registry.ccompat.dto.SchemaInfo;
 import io.apicurio.registry.ccompat.dto.SubjectVersion;
 import io.apicurio.registry.ccompat.rest.v7.SchemasResource;
+import io.apicurio.registry.content.ContentHandle;
 import io.apicurio.registry.metrics.health.liveness.ResponseErrorLivenessCheck;
 import io.apicurio.registry.metrics.health.readiness.ResponseTimeoutReadinessCheck;
+import io.apicurio.registry.storage.ArtifactNotFoundException;
+import io.apicurio.registry.storage.dto.ArtifactMetaDataDto;
+import io.apicurio.registry.storage.dto.ArtifactReferenceDto;
+import io.apicurio.registry.storage.dto.ContentWrapperDto;
+import io.apicurio.registry.storage.dto.StoredArtifactDto;
 import io.apicurio.registry.types.ArtifactType;
-
+import io.apicurio.registry.util.ArtifactTypeUtil;
 import jakarta.interceptor.Interceptors;
+
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static io.apicurio.registry.storage.RegistryStorage.ArtifactRetrievalBehavior.DEFAULT;
 
 /**
  * @author Carles Arnal
@@ -41,8 +52,24 @@ public class SchemasResourceImpl extends AbstractResource implements SchemasReso
     @Override
     @Authorized(style = AuthorizedStyle.GlobalId, level = AuthorizedLevel.Read)
     public SchemaInfo getSchema(int id, String subject, String groupId) {
-        //subject is not used since contexts are not supported
-        return facade.getSchemaById(id);
+        ContentHandle contentHandle;
+        List<ArtifactReferenceDto> references;
+        //FIXME simplify logic
+        if (cconfig.legacyIdModeEnabled.get()) {
+            StoredArtifactDto artifactVersion = storage.getArtifactVersion(id);
+            contentHandle = artifactVersion.getContent();
+            references = artifactVersion.getReferences();
+        } else {
+            ContentWrapperDto contentWrapper = storage.getArtifactByContentId(id);
+            contentHandle = storage.getArtifactByContentId(id).getContent();
+            references = contentWrapper.getReferences();
+            List<ArtifactMetaDataDto> artifacts = storage.getArtifactVersionsByContentId(id);
+            if (artifacts == null || artifacts.isEmpty()) {
+                //the contentId points to an orphaned content
+                throw new ArtifactNotFoundException("ContentId: " + id);
+            }
+        }
+        return converter.convert(contentHandle, ArtifactTypeUtil.determineArtifactType(contentHandle, null, null, storage.resolveReferences(references), factory.getAllArtifactTypes()), references);
     }
 
     @Override
@@ -55,6 +82,15 @@ public class SchemasResourceImpl extends AbstractResource implements SchemasReso
     @Authorized(style = AuthorizedStyle.GlobalId, level = AuthorizedLevel.Read)
     public List<SubjectVersion> getSubjectVersions(int id, Boolean fdeleted) {
         boolean deleted = fdeleted != null && fdeleted;
-        return facade.getSubjectVersions(id, deleted);
+        if (cconfig.legacyIdModeEnabled.get()) {
+            ArtifactMetaDataDto artifactMetaData = storage.getArtifactMetaData(id);
+            return Collections.singletonList(converter.convert(artifactMetaData.getId(), artifactMetaData.getVersionId()));
+        }
+
+        return storage.getArtifactVersionsByContentId(id)
+                .stream()
+                .filter(artifactMetaData -> deleted || isArtifactActive(artifactMetaData.getId(), artifactMetaData.getGroupId(), DEFAULT))
+                .map(artifactMetaData -> converter.convert(artifactMetaData.getId(), artifactMetaData.getVersionId()))
+                .collect(Collectors.toList());
     }
 }
