+++ app/src/main/java/io/apicurio/registry/ccompat/rest/v7/impl/SubjectsResourceImpl.java
@@ -23,14 +23,31 @@ import io.apicurio.registry.auth.AuthorizedLevel;
 import io.apicurio.registry.auth.AuthorizedStyle;
 import io.apicurio.registry.ccompat.dto.Schema;
 import io.apicurio.registry.ccompat.dto.SchemaInfo;
+import io.apicurio.registry.ccompat.rest.error.SchemaNotFoundException;
+import io.apicurio.registry.ccompat.rest.error.SubjectNotSoftDeletedException;
+import io.apicurio.registry.ccompat.rest.error.SubjectSoftDeletedException;
 import io.apicurio.registry.ccompat.rest.v7.SubjectsResource;
 import io.apicurio.registry.metrics.health.liveness.ResponseErrorLivenessCheck;
 import io.apicurio.registry.metrics.health.readiness.ResponseTimeoutReadinessCheck;
-
+import io.apicurio.registry.storage.ArtifactNotFoundException;
+import io.apicurio.registry.storage.InvalidArtifactStateException;
+import io.apicurio.registry.storage.dto.ArtifactVersionMetaDataDto;
+import io.apicurio.registry.storage.dto.OrderBy;
+import io.apicurio.registry.storage.dto.OrderDirection;
+import io.apicurio.registry.storage.dto.SearchFilter;
+import io.apicurio.registry.storage.dto.SearchedArtifactDto;
+import io.apicurio.registry.storage.dto.StoredArtifactDto;
+import io.apicurio.registry.types.ArtifactState;
+import io.apicurio.registry.util.VersionUtil;
 import jakarta.interceptor.Interceptors;
+
+import java.util.Collections;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import static io.apicurio.common.apps.logging.audit.AuditingConstants.KEY_ARTIFACT_ID;
+import static io.apicurio.registry.storage.RegistryStorage.ArtifactRetrievalBehavior.DEFAULT;
 
 /**
  * @author Carles Arnal
@@ -44,7 +61,7 @@ public class SubjectsResourceImpl extends AbstractResource implements SubjectsRe
     public List<String> listSubjects(String subjectPrefix, Boolean deleted, String groupId) {
         //Since contexts are not supported, subjectPrefix is not used
         final boolean fdeleted = deleted == null ? Boolean.FALSE : deleted;
-        return facade.getSubjects(fdeleted, groupId);
+        return storage.searchArtifacts(Set.of(SearchFilter.ofGroup(groupId)), OrderBy.createdOn, OrderDirection.asc, 0, cconfig.maxSubjects.get()).getArtifacts().stream().filter(searchedArtifactDto -> isCcompatManagedType(searchedArtifactDto.getType()) && shouldFilterState(fdeleted, searchedArtifactDto.getState())).map(SearchedArtifactDto::getId).collect(Collectors.toList());
     }
 
     @Override
@@ -53,7 +70,23 @@ public class SubjectsResourceImpl extends AbstractResource implements SubjectsRe
         final boolean fnormalize = normalize == null ? Boolean.FALSE : normalize;
         final boolean fdeleted = deleted == null ? Boolean.FALSE : deleted;
 
-        return facade.getSchemaNormalize(subject, request, fnormalize, groupId, fdeleted);
+        if (doesArtifactExist(subject, groupId)) {
+            try {
+                ArtifactVersionMetaDataDto amd;
+                amd = lookupSchema(groupId, subject, request.getSchema(), request.getReferences(), request.getSchemaType(), fnormalize);
+                if (amd.getState() != ArtifactState.DISABLED || fdeleted) {
+                    StoredArtifactDto storedArtifact = storage.getArtifactVersion(groupId, subject, amd.getVersion());
+                    return converter.convert(subject, storedArtifact);
+                } else {
+                    throw new SchemaNotFoundException(String.format("The given schema does not match any schema under the subject %s", subject));
+                }
+            } catch (ArtifactNotFoundException anf) {
+                throw new SchemaNotFoundException(String.format("The given schema does not match any schema under the subject %s", subject));
+            }
+        } else {
+            //If the artifact does not exist there is no need for looking up the schema, just fail.
+            throw new ArtifactNotFoundException(groupId, subject);
+        }
     }
 
     @Override
@@ -61,6 +94,36 @@ public class SubjectsResourceImpl extends AbstractResource implements SubjectsRe
     @Authorized(style = AuthorizedStyle.ArtifactOnly, level = AuthorizedLevel.Write)
     public List<Integer> deleteSubject(String subject, Boolean permanent, String groupId) throws Exception {
         final boolean fpermanent = permanent == null ? Boolean.FALSE : permanent;
-        return facade.deleteSubject(subject, fpermanent, groupId);
+        //FIXME simplify logic
+        if (fpermanent) {
+            if (isArtifactActive(subject, groupId, DEFAULT)) {
+                throw new SubjectNotSoftDeletedException(String.format("Subject %s must be soft deleted first", subject));
+            } else {
+                return storage.deleteArtifact(groupId, subject).stream().map(VersionUtil::toInteger).map(converter::convertUnsigned).collect(Collectors.toList());
+            }
+
+        } else if (isArtifactActive(subject, groupId, DEFAULT)) {
+
+            List<String> deletedVersions = storage.getArtifactVersions(groupId, subject);
+
+            try {
+                deletedVersions.forEach(version -> storage.updateArtifactState(groupId, subject, version, ArtifactState.DISABLED));
+            } catch (InvalidArtifactStateException ignored) {
+                log.warn("Invalid artifact state transition", ignored);
+            }
+
+            return deletedVersions.stream().map(VersionUtil::toLong).map(converter::convertUnsigned).sorted().collect(Collectors.toList());
+
+        } else {
+            if (storage.isArtifactExists(groupId, subject)) {
+                if (isArtifactActive(subject, groupId, DEFAULT)) {
+                    return storage.getArtifactVersions(groupId, subject).stream().map(VersionUtil::toLong).map(converter::convertUnsigned).sorted().collect(Collectors.toList());
+                } else {
+                    throw new SubjectSoftDeletedException(String.format("Subject %s is in soft deleted state.", subject));
+                }
+            } else {
+                return Collections.emptyList();
+            }
+        }
     }
 }
